extracted createFacilityDto
